Update style.scss
Adjust order of objects, mainly. Each object check for a `$use-objectname` variable, all of which are set to false in _defaults.scss and overridden in _vars.scss, so if including the objects in such a granular fashion is necessary. I can definitely see the value in splitting out the base and generic files, as they are not tied to variables like objects are.